extrayendo el creador de txt
