+++ app/templates/components/mostrador-de-minuta.hbs
@@ -12,7 +12,7 @@
 <div class="card" >
   <div class="card-content">
     <span>
-      <p class="text-viewer">{{textoMinuta}}</p>
+      <p class="text-viewer crop">{{textoMinuta}}</p>
     </span>
   </div>
 </div>
